Gil - moved stdDevs to utils/math
