+++ src/main/java/frc/robot/poseestimator/WPILibPoseEstimator/WPILibPoseEstimatorConstants.java
@@ -2,7 +2,7 @@ package frc.robot.poseestimator.WPILibPoseEstimator;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.poseestimator.helpers.StandardDeviations2D;
+import frc.utils.math.StandardDeviations2D;
 import frc.robot.vision.data.AprilTagVisionData;
 
 import java.util.function.Function;
