+++ src/main/java/frc/utils/math/StandardDeviations2D.java
@@ -1,10 +1,9 @@
-package frc.robot.poseestimator.helpers;
+package frc.utils.math;
 
 import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
-import frc.robot.poseestimator.Pose2dComponentsValue;
 
 public record StandardDeviations2D(double xStandardDeviations, double yStandardDeviations, double angleStandardDeviations) {
